+++ src/visitors/http/get-api-properties.js
@@ -6,6 +6,7 @@ let unexpress = require('./un-express-route')
 module.exports = function getApiProperties (arc) {
   return {
     StageName: 'staging',
+    OpenApiVersion: '3.0.1',
     DefinitionBody: getOpenApi(arc),
     EndpointConfiguration: 'REGIONAL',
     BinaryMediaTypes: [ '*~1*' ], // wat